code_93_reader => typescript
- also make integration tests significantly more verbose, as failures in
  a test do not cause any useful indication of exactly which test failed,
  or how it failed.

+++ src/reader/code_39_reader.ts
@@ -131,7 +131,7 @@ class Code39Reader extends BarcodeReader {
         let decodedChar;
         let lastStart: number;
         do {
-            counters = this._toCounters(nextStart, counters);
+            counters = this._toCounters(nextStart, counters) as Uint16Array;
             const pattern = this._toPattern(counters);
             if (pattern < 0) {
                 return null;
